requirements.txt
geopy==2.2.0
python-dotenv==0.20.0
requests==2.28.1
pydantic==1.9.1